Rename cloudeagle to ceagle
I am to lazy to write every time cloudeagle so let's keep shorter version
