+++ setup.cfg
@@ -18,11 +18,11 @@ classifier =
 
 [files]
 packages =
-    cloudeagle
+    ceagle
 
 [entry_points]
 console_scripts =
-    cloudeagle = cloudeagle.app:main
+    ceagle = ceagle.app:main
 
 [global]
 setup-hooks =
