+++ tox.ini
@@ -29,7 +29,7 @@ commands = flake8
 distribute = false
 
 [testenv:cover]
-commands = py.test --cov=cloudeagle tests/unit/ --cov-report=html
+commands = py.test --cov=ceagle tests/unit/ --cov-report=html
 
 [testenv:py34]
 basepython = python3.4
